backend/requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-nmap==0.7.1
dnspython==2.6.1
python-whois==0.9.5
requests==2.31.0
python-dotenv==1.0.1